% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_exists.R
\name{gd_exists}
\alias{gd_exists}
\title{Check if folder exists}
\usage{
gd_exists(googledriveFolder, rootid)
}
\arguments{
\item{googledriveFolder}{Character String. Path to folder on Google Drive in which to push data.}

\item{rootid}{Drive-id. Root id on google drive in which \code{googleDriveFolder} is nested}
}
\value{
A list of 2 items. If \code{googleDriveFolder} is present, its \code{id} is returned with \code{isfound} = T.
If  \code{googleDriveFolder} is not present then the \code{id} of the parent folder is returned with \code{isfound} = F
\item{id}{Drive-id for either the parent or child folder}
\item{isfound}{Boolean. Absence (F) or presence (T) of the target folder}
}
\description{
Checks to see if target folder is already present.
}
\examples{

\dontrun{
Check if "Testing/SampleRun" folder has already been created.
gd_exists("Testing/SampleRun",atlantisdrive::rootid)

}



}
\seealso{
Other atlantisdrive functions: 
\code{\link{get_dirs}()},
\code{\link{get_files}()},
\code{\link{list_atlantisom_files}()},
\code{\link{list_core_files}()},
\code{\link{pull_from_drive}()},
\code{\link{push_to_drive}()}
}
\concept{atlantisdrive functions}
