% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dirs.R
\name{get_dirs}
\alias{get_dirs}
\title{Get list of directories on google drive}
\usage{
get_dirs(rootid = atlantisdrive::rootid)
}
\arguments{
\item{rootid}{Drive-id. Atlantis root id on google drive. (Default id for NEFSC is bundled in this package}
}
\value{
A tibble containing folder names, id, etc
}
\description{
Lists all folders under the root google drive folder
}
\examples{
\dontrun{
# get list of directories under the root
get_dirs()
}

}
\seealso{
Other atlantisdrive functions: \code{\link{get_files}},
  \code{\link{list_atlantisom_files}},
  \code{\link{list_core_files}},
  \code{\link{pull_from_drive}},
  \code{\link{push_to_drive}}
}
\concept{atlantisdrive functions}
