% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_list.R
\name{get_file_list}
\alias{get_file_list}
\title{Get list of files with a directory on Google drive}
\usage{
get_file_list(rootid = atlantisdrive::rootid, targetDir = NULL)
}
\arguments{
\item{rootid}{Drive-id. The id to the root Atlantis Folder}

\item{targetDir}{Character. Path to a directory}
}
\value{
A tibble containing file name details
\item{name}{Folder/filename}
\item{id}{Drive id of folder/filename}
\item{drive_resource}{list of 33 objects defind in [googledrive](drive_ls)}
}
\description{
Lists all files under a directory on Google drive folder
}
\examples{
\dontrun{
# get list of files in the Scenario folder under the root
get_files(targetDir = "Scenario")
get_files(targetDir = "Scen")
get_files(targetDir = "ario")

}

}
\seealso{
Other atlantisdrive functions: 
\code{\link{gd_exists}()},
\code{\link{list_atlantisom_files}()},
\code{\link{list_core_files}()},
\code{\link{pull_from_drive}()},
\code{\link{push_to_drive}()}
}
\concept{atlantisdrive functions}
