% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Create a list of Core files used in Atlantis}
\usage{
get_files(path = here::here(), filesToInclude = c("txt", "nc", "ts",
  "xml", "bgm"))
}
\arguments{
\item{path}{Character string. Path to location of atlantis output files}

\item{filesToInclude}{Character vector. Types of files to copy (Default = ("txt","nc","ts","xml","bgm"))}
}
\value{
A character vector containing the names of all of the files to be copied to the google drive
}
\description{
Compiles a list of file names based on file extension (e.g. nc, txt, ts, xml, bgm)
}
