% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_atlantisom_files.R
\name{list_atlantisom_files}
\alias{list_atlantisom_files}
\title{List of Atlantis output files needed for atlantisom R package}
\usage{
list_atlantisom_files(scenario)
}
\arguments{
\item{scenario}{Character vector. User specified model name/scenario portion of atlantis output filenames}
}
\value{
A character vector containing the names of all of the files to be copied to the google drive
}
\description{
Compiles a list of file names based on user input scenario and required atlantisom inputs
}
\examples{
\dontrun{
# create a list of atlantisom input files for Norweigan Barents Sea model
atlantisom_outfilelist(scenario="nordic_runresults_01")

# create a list of atlantisom input files for California Current model
atlantisom_outfilelist(scenario="outputCCV3")

}

}
\seealso{
Other atlantisdrive functions: \code{\link{get_dirs}},
  \code{\link{get_files}}, \code{\link{list_core_files}},
  \code{\link{pull_from_drive}},
  \code{\link{push_to_drive}}
}
\concept{atlantisdrive functions}
