% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_core_files.R
\name{list_core_files}
\alias{list_core_files}
\title{List of Core files used in Atlantis}
\usage{
list_core_files(path = here::here(), filesToInclude = c("txt", "nc",
  "ts", "xml", "bgm"))
}
\arguments{
\item{path}{Character string. Path to location of atlantis output files}

\item{filesToInclude}{Character vector. Types of files to copy (Default = ("txt","nc","ts","xml","bgm"))}
}
\value{
A character vector containing the names of all of the files to be copied to the google drive
}
\description{
Compiles a list of file names based on file extension (e.g. nc, txt, ts, xml, bgm)
}
\examples{
\dontrun{
# create a list of all xml files in your projects "output" folder
list_core_files(path=here::here("output"),filesToInclude=c("xml"))

# create a list of all xml and ts files in your projects "test" folder
list_core_files(path=here::here("test"),filesToInclude=c("xml","ts"))

}

}
\seealso{
Other atlantisdrive functions: \code{\link{get_dirs}},
  \code{\link{get_files}}, \code{\link{pull_from_drive}},
  \code{\link{push_to_drive}}
}
\concept{atlantisdrive functions}
