% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_drive.R
\name{pull_from_drive}
\alias{pull_from_drive}
\title{Renames, moves output files from google Drive to local machine}
\usage{
pull_from_drive(id, pathToOutput = here::here("output"),
  fileList = "all", googledriveFolder, rootid = atlantisdrive::rootid)
}
\arguments{
\item{id}{Character string. Name of task/scenario/id associated with output.}

\item{fileList}{Character vector. Files that need to be pulled from Google Drive. (Default = "all")}

\item{googledriveFolder}{Character String. Name of directory on google Drive to pull.}

\item{rootid}{Drive-id. Atlantis root id on google drive. (Default id for NEFSC is bundled in this package)}

\item{filePath}{Character string. Path to local Atlantis output directory. Files will be pulled to this directtory}
}
\value{
A vector of filenames. These are the files that were not present on Google Drive and were not pulled.
}
\description{
Used for pulling archived Atlantis runs into the users workspace.
}
\seealso{
Other googldrive functions: \code{\link{push_to_drive}}
}
\concept{googldrive functions}
