% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Get list of files with a diresctory on google drive}
\usage{
get_files(rootid = atlantisdrive::rootid, targetDir)
}
\value{
A tibble containing file name, id etc.
}
\description{
Lists all files under a directory on google drive folder
}
\seealso{
Other googldrive functions: \code{\link{get_dirs}},
  \code{\link{list_core_files}},
  \code{\link{pull_from_drive}},
  \code{\link{push_to_drive}}
}
\concept{googldrive functions}
