% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Get list of files with a diresctory on google drive}
\usage{
get_files(rootid = atlantisdrive::rootid, targetDir)
}
\value{
A tibble containing file name, id etc.
}
\description{
Lists all files under a directory on google drive folder
}
\examples{
\dontrun{
# get list of files in the Scenario folder under the root
get_files(targetDir = "Scenario")
get_files(targetDir = "Scen")
get_files(targetDir = "ario")

}

}
\seealso{
Other atlantisdrive functions: \code{\link{get_dirs}},
  \code{\link{list_core_files}},
  \code{\link{pull_from_drive}},
  \code{\link{push_to_drive}}
}
\concept{atlantisdrive functions}
