% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_atlantisom_files.R
\name{list_atlantisom_files}
\alias{list_atlantisom_files}
\title{List of Atlantis output files needed for atlantisom R package}
\usage{
list_atlantisom_files(scenario)
}
\arguments{
\item{scenario}{Character vector. User specified model name/scenario portion of atlantis output filenames}
}
\value{
A character vector containing the names of all of the files to be copied to the google drive
}
\description{
Compiles a list of file names based on user input scenario and required atlantisom inputs
}
\examples{
\dontrun{
# create a list of atlantisom input files for Norweigan Barents Sea model
list_atlantisom_files(scenario="nordic_runresults_01")

# create a list of atlantisom input files for California Current model
list_atlantisom_files(scenario="outputCCV3")

}

}
\seealso{
Other atlantisdrive functions: 
\code{\link{gd_exists}()},
\code{\link{get_file_list}()},
\code{\link{list_core_files}()},
\code{\link{pull_from_drive}()},
\code{\link{push_to_drive}()}
}
\concept{atlantisdrive functions}
