% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_from_drive.R
\name{pull_from_drive}
\alias{pull_from_drive}
\title{Renames, moves output files from Google Drive to local machine}
\usage{
pull_from_drive(
  localPath = here::here(),
  fileList,
  googledriveFolder,
  rootid = atlantisdrive::rootid
)
}
\arguments{
\item{localPath}{Character string. Path to local Atlantis output directory. Files will be pulled to this directtory. Default = here::here()}

\item{fileList}{Character vector. Files name(s) that need to be pulled from Google Drive. Alternatively any character string present in file name}

\item{googledriveFolder}{Character String. Name of directory on google Drive to pull.}

\item{rootid}{Drive-id. Atlantis root id on google drive. (Default id for NEFSC is bundled in this package)}
}
\value{
Files are pulled to local drive
}
\description{
Used for pulling archived Atlantis runs into the users workspace.
}
\examples{

\dontrun{
# Pulls all xml files from the Development Folder to your current working directory
pull_from_drive(fileList="xml",googledriveFolder="Development")

# Pulls all files with the string "ATLNTS-2" present in its filename to the "output" folder in your project
pull_from_drive(localPath = here::here("output"),fileList="ATLNTS-2",googledriveFolder="Development")

# Pulls selected files from the Scenario Folder
filesToPull <- c("file1.jpg","file2.txt","file3.bmg")
pull_from_drive(fileList=filesToPull,googledriveFolder="Scenario")

}

}
\seealso{
Other atlantisdrive functions: 
\code{\link{gd_exists}()},
\code{\link{get_file_list}()},
\code{\link{list_atlantisom_files}()},
\code{\link{list_core_files}()},
\code{\link{push_to_drive}()}
}
\concept{atlantisdrive functions}
